console/scripts/requirements.txt
flask
mysql.connector
gunicorn
